+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -161,6 +161,7 @@ class GeckoEngine(
             set(value) {
                 value?.let {
                     runtime.settings.contentBlocking.categories = it.categories
+                    runtime.settings.contentBlocking.cookieBehavior = it.cookiePolicy.id
                     defaultSettings?.trackingProtectionPolicy = value
                 }
             }