+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/EngineSession.kt
@@ -6,6 +6,8 @@ package mozilla.components.concept.engine
 
 import android.graphics.Bitmap
 import androidx.annotation.CallSuper
+import mozilla.components.concept.engine.EngineSession.TrackingProtectionPolicy.CookiePolicy.ACCEPT_ALL
+import mozilla.components.concept.engine.EngineSession.TrackingProtectionPolicy.CookiePolicy.ACCEPT_NON_TRACKERS
 import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.concept.engine.media.Media
 import mozilla.components.concept.engine.media.RecordingDevice
@@ -90,8 +92,46 @@ abstract class EngineSession(
     open class TrackingProtectionPolicy internal constructor(
         val categories: Int,
         val useForPrivateSessions: Boolean = true,
-        val useForRegularSessions: Boolean = true
+        val useForRegularSessions: Boolean = true,
+        val cookiePolicy: CookiePolicy = ACCEPT_NON_TRACKERS
     ) {
+
+        /**
+         * Indicates how cookies should behave for a given [TrackingProtectionPolicy].
+         * The ids of each cookiePolicy is aligned with the GeckoView @CookieBehavior constants.
+         */
+        @Suppress("MagicNumber")
+        enum class CookiePolicy(val id: Int) {
+            /**
+             * Accept first-party and third-party cookies and site data.
+             */
+            ACCEPT_ALL(0),
+
+            /**
+             * Accept only first-party cookies and site data to block cookies which are
+             * not associated with the domain of the visited site.
+             */
+            ACCEPT_ONLY_FIRST_PARTY(1),
+
+            /**
+             * Do not store any cookies and site data.
+             */
+            ACCEPT_NONE(2),
+
+            /**
+             * Accept first-party and third-party cookies and site data only from
+             * sites previously visited in a first-party context.
+             */
+            ACCEPT_VISITED(3),
+
+            /**
+             * Accept only first-party and non-tracking third-party cookies and site data
+             * to block cookies which are not associated with the domain of the visited
+             * site set by known trackers.
+             */
+            ACCEPT_NON_TRACKERS(4)
+        }
+
         companion object {
             internal const val NONE: Int = 0
             /**
@@ -143,15 +183,16 @@ abstract class EngineSession(
             const val SAFE_BROWSING_ALL =
                 SAFE_BROWSING_MALWARE + SAFE_BROWSING_UNWANTED + SAFE_BROWSING_HARMFUL + SAFE_BROWSING_PHISHING
 
-            internal const val RECOMMENDED = AD + ANALYTICS + SOCIAL + TEST + SAFE_BROWSING_ALL
+            const val RECOMMENDED = AD + ANALYTICS + SOCIAL + TEST + SAFE_BROWSING_ALL
 
-            internal const val ALL = RECOMMENDED + CRYPTOMINING + FINGERPRINTING + CONTENT
+            const val ALL = RECOMMENDED + CRYPTOMINING + FINGERPRINTING + CONTENT
 
-            fun none() = TrackingProtectionPolicy(NONE)
+            fun none() = TrackingProtectionPolicy(NONE, cookiePolicy = ACCEPT_ALL)
 
             /**
              * Strict policy.
              * Combining the [recommended] categories plus [CRYPTOMINING], [FINGERPRINTING] and [CONTENT].
+             * With a cookiePolicy of [ACCEPT_NON_TRACKERS].
              * This is the strictest setting and may cause issues on some web sites.
              */
             fun all() = TrackingProtectionPolicyForSessionTypes(ALL)
@@ -159,11 +200,13 @@ abstract class EngineSession(
             /**
              * Recommended policy.
              * Combining the [AD], [ANALYTICS], [SOCIAL], [TEST] categories plus [SAFE_BROWSING_ALL].
+             * With a [CookiePolicy] of [ACCEPT_NON_TRACKERS].
              * This is the recommended setting.
              */
             fun recommended() = TrackingProtectionPolicyForSessionTypes(RECOMMENDED)
 
-            fun select(vararg categories: Int) = TrackingProtectionPolicyForSessionTypes(categories.sum())
+            fun select(vararg categories: Int, cookiePolicy: CookiePolicy = ACCEPT_NON_TRACKERS) =
+                TrackingProtectionPolicyForSessionTypes(categories.sum(), cookiePolicy)
         }
 
         fun contains(category: Int) = (categories and category) != 0
@@ -185,15 +228,17 @@ abstract class EngineSession(
      * should be applied to. By default, a policy will be applied to all sessions.
      */
     class TrackingProtectionPolicyForSessionTypes internal constructor(
-        categories: Int
-    ) : TrackingProtectionPolicy(categories) {
+        categories: Int,
+        cookiePolicy: CookiePolicy = ACCEPT_NON_TRACKERS
+    ) : TrackingProtectionPolicy(categories, cookiePolicy = cookiePolicy) {
         /**
          * Marks this policy to be used for private sessions only.
          */
         fun forPrivateSessionsOnly() = TrackingProtectionPolicy(
             categories,
             useForPrivateSessions = true,
-            useForRegularSessions = false
+            useForRegularSessions = false,
+            cookiePolicy = cookiePolicy
         )
 
         /**
@@ -202,7 +247,8 @@ abstract class EngineSession(
         fun forRegularSessionsOnly() = TrackingProtectionPolicy(
             categories,
             useForPrivateSessions = false,
-            useForRegularSessions = true
+            useForRegularSessions = true,
+            cookiePolicy = cookiePolicy
         )
     }